+++ tests/output/dynamic.py
@@ -6,7 +6,6 @@ from __future__ import unicode_literals
 
 import unittest
 
-import tests.output.test_lib
 from plaso.containers import events
 from plaso.formatters import interface as formatters_interface
 from plaso.formatters import manager as formatters_manager
@@ -51,7 +50,7 @@ class DynamicOutputModuleTest(test_lib.OutputModuleTestCase):
   def testHeader(self):
     """Tests the WriteHeader function."""
     output_mediator = self._CreateOutputMediator()
-    output_writer = tests.output.test_lib.TestOutputWriter()
+    output_writer = cli_test_lib.TestOutputWriter()
     output_module = dynamic.DynamicOutputModule(output_mediator)
     output_module.SetOutputWriter(output_writer)
     expected_header = (
@@ -63,7 +62,7 @@ class DynamicOutputModuleTest(test_lib.OutputModuleTestCase):
     self.assertEqual(header, expected_header)
 
     output_mediator = self._CreateOutputMediator()
-    output_writer = tests.output.test_lib.TestOutputWriter()
+    output_writer = cli_test_lib.TestOutputWriter()
     output_module = dynamic.DynamicOutputModule(output_mediator)
     output_module.SetFields([
         'date', 'time', 'message', 'hostname', 'filename', 'some_stuff'])
@@ -75,7 +74,7 @@ class DynamicOutputModuleTest(test_lib.OutputModuleTestCase):
     self.assertEqual(header, expected_header)
 
     output_mediator = self._CreateOutputMediator()
-    output_writer = tests.output.test_lib.TestOutputWriter()
+    output_writer = cli_test_lib.TestOutputWriter()
     output_module = dynamic.DynamicOutputModule(output_mediator)
     output_module.SetFields([
         'date', 'time', 'message', 'hostname', 'filename', 'some_stuff'])
@@ -95,7 +94,7 @@ class DynamicOutputModuleTest(test_lib.OutputModuleTestCase):
     event_object = TestEvent()
 
     output_mediator = self._CreateOutputMediator()
-    output_writer = tests.output.test_lib.TestOutputWriter()
+    output_writer = cli_test_lib.TestOutputWriter()
     output_module = dynamic.DynamicOutputModule(output_mediator)
     output_module.SetFields([
         'date', 'time', 'timezone', 'macb', 'source', 'sourcetype',
@@ -121,7 +120,7 @@ class DynamicOutputModuleTest(test_lib.OutputModuleTestCase):
     self.assertEqual(event_body, expected_event_body)
 
     output_mediator = self._CreateOutputMediator()
-    output_writer = tests.output.test_lib.TestOutputWriter()
+    output_writer = cli_test_lib.TestOutputWriter()
     output_module = dynamic.DynamicOutputModule(output_mediator)
     output_module.SetFields([
         'datetime', 'nonsense', 'hostname', 'message'])
