+++ tests/output/l2t_csv.py
@@ -6,6 +6,7 @@ from __future__ import unicode_literals
 
 import unittest
 
+import tests.cli.test_lib
 from plaso.containers import events
 from plaso.formatters import interface as formatters_interface
 from plaso.formatters import manager as formatters_manager
@@ -13,6 +14,7 @@ from plaso.lib import definitions
 from plaso.lib import timelib
 from plaso.output import l2t_csv
 
+from tests.cli import test_lib as cli_test_lib
 from tests.output import test_lib
 
 
@@ -50,7 +52,7 @@ class L2TCSVTest(test_lib.OutputModuleTestCase):
   def setUp(self):
     """Makes preparations before running an individual test."""
     output_mediator = self._CreateOutputMediator()
-    self._output_writer = test_lib.TestOutputWriter()
+    self._output_writer = cli_test_lib.TestOutputWriter()
     self._formatter = l2t_csv.L2TCSVOutputModule(output_mediator)
     self._formatter.SetOutputWriter(self._output_writer)
