+++ tests/output/tln.py
@@ -9,6 +9,7 @@ import unittest
 from plaso.formatters import manager as formatters_manager
 from plaso.output import tln
 
+from tests.cli import test_lib as cli_test_lib
 from tests.output import test_lib
 
 
@@ -17,7 +18,7 @@ class TLNOutputModuleTest(test_lib.OutputModuleTestCase):
 
   def setUp(self):
     """Makes preparations before running an individual test."""
-    self._output_writer = test_lib.TestOutputWriter()
+    self._output_writer = cli_test_lib.TestOutputWriter()
     output_mediator = self._CreateOutputMediator()
     self._output_module = tln.TLNOutputModule(output_mediator)
     self._output_module.SetOutputWriter(self._output_writer)
@@ -59,7 +60,7 @@ class L2TTLNOutputModuleTest(test_lib.OutputModuleTestCase):
   def setUp(self):
     """Makes preparations before running an individual test."""
     output_mediator = self._CreateOutputMediator()
-    self._output_writer = test_lib.TestOutputWriter()
+    self._output_writer = cli_test_lib.TestOutputWriter()
     self._output_module = tln.L2TTLNOutputModule(output_mediator)
     self._output_module.SetOutputWriter(self._output_writer)
     self._event_object = test_lib.TestEventObject()
